COMP: Fix -Wsign-compare in vtkFSSurfaceAnnotationReader.cxx
This commit fixes warning like the following:

/tmp/SlicerFreeSurfer/FreeSurfer/vtkFSSurfaceAnnotationReader.cxx:1026:60: warning: comparison of integer expressions of different signedness: ‘int’ and ‘std::vector<std::vector<int, std::allocator<int> > >::size_type’ {aka ‘long unsigned int’} [-Wsign-compare]
 1026 |   for (int colorTableEntryIndex = 0; (colorTableEntryIndex < colors.size());
      |                                       ~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~

+++ FreeSurfer/vtkFSSurfaceAnnotationReader.cxx
@@ -997,7 +997,7 @@ int vtkFSSurfaceAnnotationReader::GenerateColorTable(
 
     bool found = false;
     for (int colorTableEntryIndex = 0;
-      (found == false) && (colorTableEntryIndex < colors.size());
+      (found == false) && (colorTableEntryIndex < static_cast<int>(colors.size()));
       colorTableEntryIndex++)
     {
       if (r == colors[colorTableEntryIndex][0] &&
@@ -1023,7 +1023,7 @@ int vtkFSSurfaceAnnotationReader::GenerateColorTable(
   int numColorTableEntries = colors.size();
   int** colorTableRGBs = (int**)calloc(numColorTableEntries, sizeof(int*));
   char** colorTableNames = (char**)calloc(numColorTableEntries, sizeof(char*));
-  for (int colorTableEntryIndex = 0; (colorTableEntryIndex < colors.size());
+  for (int colorTableEntryIndex = 0; (colorTableEntryIndex < static_cast<int>(colors.size()));
     colorTableEntryIndex++)
   {
     int* color = (int*)calloc(3, sizeof(int));
